requirements.txt
pyspark>=3.0.0
pandas>=1.3.0
numpy>=1.21.0
kafka-python>=2.0.2
pyarrow>=10.0.1
